# bureau/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for the Bureau MCP Server.",
        "title": "Bureau API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/admin": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Get a list of all admins.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/admin"
                ],
                "summary": "List all admins",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit the number of results",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset for pagination",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.PaginatedResponse-github_com_BrunoPolaski_bureau_internal_infra_controllers_dto_AdminDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Creates a new admin with the provided details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/admin"
                ],
                "summary": "Create a new admin",
                "parameters": [
                    {
                        "description": "Personal Information",
                        "name": "personalInformation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_BrunoPolaski_bureau_internal_infra_controllers_request.PersonalInformationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.AdminDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Delete an admin by its ID.",
                "tags": [
                    "/admin"
                ],
                "summary": "Delete an admin",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Admin ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    }
                }
            }
        },
        "/admin/{id}": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Get an admin by its ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/admin"
                ],
                "summary": "Get an admin by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Admin ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.AdminDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    }
                }
            }
        },
...
        "/auth/register": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates a new user with the provided details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/auth"
                ],
                "summary": "Create a new user",
                "parameters": [
                    {
                        "description": "User to create",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.UserDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.UserDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    }
                }
            }
        },
        "/auth/signin": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": [],
                        "BasicAuth": []
                    }
                ],
                "description": "Signs in with the provided credentials",
                "tags": [
                    "/auth"
                ],
                "summary": "Sign in",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.AdminDTO": {
            "type": "object",
            "required": [
                "personal_information"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "personal_information": {
                    "$ref": "#/definitions/github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.PersonalInformationDTO"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.AnalystDTO": {
            "type": "object",
            "required": [
                "personal_information"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "personal_information": {
                    "$ref": "#/definitions/github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.PersonalInformationDTO"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.PaginatedResponse-github_com_BrunoPolaski_bureau_internal_infra_controllers_dto_AdminDTO": {
            "type": "object",
            "properties": {
                "items": {
                    "description": "List of items in the current page",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.AdminDTO"
                    }
                },
                "total": {
                    "description": "Total number of items available",
                    "type": "integer"
                }
            }
        },
...
        "github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.PersonAddressDTO": {
            "type": "object",
            "properties": {
                "address_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.PersonDTO": {
            "type": "object",
            "required": [
                "personal_information"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "credit_accounts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.CreditAccountDTO"
                    }
                },
                "credit_inquiries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.CreditInquiryDTO"
                    }
                },
                "credit_score": {
                    "$ref": "#/definitions/github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.CreditScoreDTO"
                },
                "credit_score_id": {
                    "type": "integer"
                },
                "data_sources": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.DataSourceDTO"
                    }
                },
                "debts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.DebtDTO"
                    }
                },
                "employment_records": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.EmploymentRecordDTO"
                    }
                },
                "financial_profile": {
                    "$ref": "#/definitions/github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.FinancialProfileDTO"
                },
                "financial_profile_id": {
                    "type": "integer"
                },
                "fraud_alerts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.FraudAlertDTO"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "last_verified_at": {
                    "type": "string"
                },
                "legal_records": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.LegalRecordDTO"
                    }
                },
                "negative_records": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.NegativeRecordDTO"
                    }
                },
                "payment_histories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.PaymentHistoryDTO"
                    }
                },
                "personal_information": {
                    "$ref": "#/definitions/github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.PersonalInformationDTO"
                },
                "personal_information_id": {
                    "type": "integer"
                },
                "related_persons": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.PersonRelationshipDTO"
                    }
                },
                "risk_assessments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.RiskAssessmentDTO"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.PersonalInformationDTO": {
            "type": "object",
            "required": [
                "full_name"
            ],
            "properties": {
                "addresses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.PersonAddressDTO"
                    }
                },
                "alternative_email": {
                    "type": "string"
                },
                "birth_date": {
                    "type": "string"
                },
                "document": {
                    "type": "string",
                    "example": "12345678909"
                },
                "documents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.PersonDocumentDTO"
                    }
                },
                "email": {
                    "type": "string"
                },
                "email_verified": {
                    "type": "boolean"
                },
                "full_name": {
                    "type": "string",
                    "example": "João da Silva"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "marital_status": {
                    "type": "string"
                },
                "mother_name": {
                    "type": "string"
                },
                "nationality": {
                    "type": "string"
                },
                "phone_verified": {
                    "type": "boolean"
                },
                "primary_phone": {
                    "type": "string",
                    "example": "11999999999"
                },
                "profile_photo": {
                    "$ref": "#/definitions/github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.FileDTO"
                },
                "rg": {
                    "type": "string"
                },
                "rg_issue_date": {
                    "type": "string"
                },
                "rg_issuer": {
                    "type": "string"
                },
                "secondary_phone": {
                    "type": "string"
                },
                "voter_id": {
                    "type": "string"
                },
                "work_card": {
                    "type": "string"
                }
            }
        },
...
        "github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.UserDTO": {
            "type": "object",
            "required": [
                "email",
                "password",
                "user_type"
            ],
            "properties": {
                "additional_information_required": {
                    "type": "boolean",
                    "example": false
                },
                "admin": {
                    "$ref": "#/definitions/github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.AdminDTO"
                },
                "analyst": {
                    "$ref": "#/definitions/github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.AnalystDTO"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "id": {
                    "type": "integer"
                },
                "password": {
                    "type": "string",
                    "example": "P@ssw0rd!"
                },
                "person": {
                    "$ref": "#/definitions/github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.PersonDTO"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_type": {
                    "type": "string",
                    "example": "person | owner | admin | analyst"
                }
            }
        },
        "github_com_BrunoPolaski_bureau_internal_infra_controllers_request.FileRequest": {
            "type": "object",
            "required": [
                "original_name",
                "url"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "mime_type": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "original_name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "type": "string",
                    "example": "https://example.com/file.jpg"
                }
            }
        },
        "github_com_BrunoPolaski_bureau_internal_infra_controllers_request.PersonAddressRequest": {
            "type": "object",
            "required": [
                "address_id"
            ],
            "properties": {
                "address_id": {
                    "type": "integer"
                }
            }
        },
        "github_com_BrunoPolaski_bureau_internal_infra_controllers_request.PersonDocumentRequest": {
            "type": "object",
            "required": [
                "document_type"
            ],
            "properties": {
                "document_type": {
                    "type": "string"
                },
                "expiration_date": {
                    "type": "string"
                },
                "file": {
                    "$ref": "#/definitions/github_com_BrunoPolaski_bureau_internal_infra_controllers_request.FileRequest"
                },
                "id": {
                    "type": "integer"
                },
                "is_verified": {
                    "type": "boolean"
                },
                "verified_at": {
                    "type": "string"
                },
                "verified_by": {
                    "type": "string"
                }
            }
        },
        "github_com_BrunoPolaski_bureau_internal_infra_controllers_request.PersonalInformationRequest": {
            "type": "object",
            "required": [
                "document",
                "full_name"
            ],
            "properties": {
                "addresses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_BrunoPolaski_bureau_internal_infra_controllers_request.PersonAddressRequest"
                    }
                },
                "alternative_email": {
                    "type": "string"
                },
                "biometric_validated": {
                    "type": "boolean"
                },
                "birth_date": {
                    "type": "string"
                },
                "document": {
                    "type": "string",
                    "example": "12345678909"
                },
                "document_validated": {
                    "type": "boolean"
                },
                "documents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_BrunoPolaski_bureau_internal_infra_controllers_request.PersonDocumentRequest"
                    }
                },
                "email": {
                    "type": "string"
                },
                "email_verified": {
                    "type": "boolean"
                },
                "full_name": {
                    "type": "string",
                    "example": "João da Silva"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "marital_status": {
                    "type": "string"
                },
                "mother_name": {
                    "type": "string"
                },
                "nationality": {
                    "type": "string"
                },
                "phone_verified": {
                    "type": "boolean"
                },
                "primary_phone": {
                    "type": "string",
                    "example": "11999999999"
                },
                "profile_photo": {
                    "$ref": "#/definitions/github_com_BrunoPolaski_bureau_internal_infra_controllers_request.FileRequest"
                },
                "receita_federal_status": {
                    "type": "string"
                },
                "rg": {
                    "type": "string"
                },
                "rg_issue_date": {
                    "type": "string"
                },
                "rg_issuer": {
                    "type": "string"
                },
                "secondary_phone": {
                    "type": "string"
                },
                "voter_id": {
                    "type": "string"
                },
                "work_card": {
                    "type": "string"
                }
            }
        },
        "rest_err.Causes": {
            "type": "object",
            "properties": {
                "field": {
                    "description": "Field or parameter that caused the error",
                    "type": "string",
                    "example": "email"
                },
                "message": {
                    "description": "Description of the cause",
                    "type": "string",
                    "example": "invalid email address"
                }
            }
        },
        "rest_err.RestErr": {
            "type": "object",
            "properties": {
                "causes": {
                    "description": "Detailed error causes, most common for json field validation errors",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest_err.Causes"
                    }
                },
                "code": {
                    "description": "HTTP status code",
                    "type": "integer",
                    "example": 400
                },
                "error": {
                    "type": "string",
                    "example": "bad request"
                },
                "message": {
                    "description": "Human readable message",
                    "type": "string",
                    "example": "invalid request parameters"
                },
                "timestamp": {
                    "description": "When the error occurred",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "x-api-key",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "CookieAuth": {
            "type": "apiKey",
            "name": "Sid",
            "in": "cookie"
        }
    }
}